test(dbus): fix cross-test gi.repository pollution
The D-Bus activation tests imported vocalinux modules (and TrayIndicator) at
module level. Because the file is collected early, that froze
tray_indicator.GLib against a stale gi.repository mock; other test files swap
sys.modules["gi.repository"] at runtime, so later GLib-identity assertions
(e.g. GLib.SOURCE_REMOVE in test_suspend_handler) and single-instance tests
failed depending on collection order.

Make all vocalinux imports in test_dbus_activation.py lazy, mock
single_instance via sys.modules instead of patching acquire_lock, and move the
_setup_keyboard_shortcuts gate tests to test_tray_indicator_ext.py, where
tray_indicator is imported late (after the gi.repository swaps).
